api/Python/import_metadata/requirements.txt
openslide-python==1.1.2
pylibdmtx==0.1.9
requests==2.25.1